Changed organization of classes in files
