+++ shortcut_composer/templates/pie_menu_utils/settings_gui/pie_settings.py
@@ -16,8 +16,7 @@ from api_krita import Krita
 from api_krita.pyqt import AnimatedWidget, BaseWidget, SafeConfirmButton
 from config_system.ui import ConfigFormWidget, ConfigSpinBox
 from composer_utils import Config
-from ..pie_style import PieStyle
-from ..pie_config import PieConfig
+from templates.pie_menu_utils import PieStyle, PieConfig
 
 
 class PieSettings(AnimatedWidget, BaseWidget):
